requirements.txt
openai>=0.27.0
python-dotenv>=1.0.0
pypdf>=3.8.0
pytest>=7.0.0
tqdm>=4.0.0
